+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Deprecated methods removed:
 - `MoneyMismatchException` now has explicit `CurrencyMismatchException` and `ContextMismatchException` subclasses
 - New exception: `MoneyFormatException`, thrown by `MoneyFormatter::format()`
 - New methods in `MoneyBag`: `isZero()`, `multipliedBy()`, `dividedBy()`
+- New methods in `RationalMoney`: `min()`, `max()`, `sum()`
 - Support for custom PSR-16 cache implementations in `CachedProvider`
 
 ## [0.12.3](https://github.com/brick/money/releases/tag/0.12.3) - 2026-03-23